Major Pattern class refactor
Lots of simplifications: Event class has ceased to be. Patterns are
constructed from values only (or a block that yields values), and only
when they are consumed on #each_event is when they acquire an onset and
duration based on Pattern's delta values.

Another important change: Pattern#each_event accepts a +cycle+ numeric
argument that specifies when to start enumerating values from.  This
moves complexity for "forwarding" and "playing" enumerators from Stream
to Pattern in a more sensible way.  There are still some rough edges,
but it is much more manageable now.